Add support for Home and End keys
The [WAI ARIA Practices 1.1][listbox-keyboard] outlines optional support
for [`Home` and `End` keys][keys] to jump to the top and bottom of the
listbox, respectively.

[listbox-keyboard]: https://www.w3.org/WAI/ARIA/apg/patterns/combobox/#listbox-popup-keyboard-interaction
[keys]: https://www.w3.org/TR/uievents-key/#keys-navigation

Configure: Home and End navigation
---

By default, continue to ignore `Home` and `End` navigation keys.

When `optionalNavigationKeys: ["Home", "End"]` is passed as a
configuration object, treat them as valid navigation keys.